requirements.txt
Flask==2.3.3
python-dotenv==1.1.0
groq==0.31.0
